editted filter
